Update material.css
This fixes an issue that is on the Discussions List, when a User hovers over one, a Drop Down Button is then visible and was misaligned and then later hiding the Reply count. This will replace this issue with a fix that move the button just a few more Pixels to the Right while still allowing you to use it with a Material look and feel.